Updates to let HBMDramSim & DramSim coexist
