--- DRAMSim.h
@@ -1,68 +1,0 @@
-/*********************************************************************************
-*  Copyright (c) 2010-2011, Elliott Cooper-Balis
-*                             Paul Rosenfeld
-*                             Bruce Jacob
-*                             University of Maryland 
-*                             dramninjas [at] gmail [dot] com
-*  All rights reserved.
-*  
-*  Redistribution and use in source and binary forms, with or without
-*  modification, are permitted provided that the following conditions are met:
-*  
-*     * Redistributions of source code must retain the above copyright notice,
-*        this list of conditions and the following disclaimer.
-*  
-*     * Redistributions in binary form must reproduce the above copyright notice,
-*        this list of conditions and the following disclaimer in the documentation
-*        and/or other materials provided with the distribution.
-*  
-*  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-*  ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-*  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-*  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-*  FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-*  DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-*  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-*  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-*  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-*  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*********************************************************************************/
-
-
-#ifndef DRAMSIM_H
-#define DRAMSIM_H
-
-#include "Callback.h"
-#include <string>
-#include "Stats.h"
-
-using std::string;
-
-namespace DRAMSim
-{
-  class MultiChannelMemorySystem
-  {
-    public:
-      bool addTransaction(bool isWrite, uint64_t addr);
-      void update();
-      void printStats(bool finalStats);
-      bool willAcceptTransaction();
-      bool willAcceptTransaction(uint64_t addr);
-
-      void RegisterCallbacks(TransactionCompleteCB *readDone,
-                             TransactionCompleteCB *writeDone,
-                             void (*reportPower)(double bgpower,
-                                                 double burstpower,
-                                                 double refreshpower,
-                                                 double actprepower));
-
-      // SST Statistics
-      bool getStats( double *stat, DSIM_STAT metric );
-      bool getStats( uint64_t *stat, DSIM_STAT metric );
-  };
-
-  MultiChannelMemorySystem *getMemorySystemInstance(const string &dev, const string &sys, 
-      const string &pwd, unsigned megsOfMemory);
-}
-
-#endif
