+++ HBMDRAMSim.h
@@ -2,20 +2,20 @@
 *  Copyright (c) 2010-2011, Elliott Cooper-Balis
 *                             Paul Rosenfeld
 *                             Bruce Jacob
-*                             University of Maryland
+*                             University of Maryland 
 *                             dramninjas [at] gmail [dot] com
 *  All rights reserved.
-*
+*  
 *  Redistribution and use in source and binary forms, with or without
 *  modification, are permitted provided that the following conditions are met:
-*
+*  
 *     * Redistributions of source code must retain the above copyright notice,
 *        this list of conditions and the following disclaimer.
-*
+*  
 *     * Redistributions in binary form must reproduce the above copyright notice,
 *        this list of conditions and the following disclaimer in the documentation
 *        and/or other materials provided with the distribution.
-*
+*  
 *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
 *  ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 *  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -28,15 +28,41 @@
 *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************************/
 
-/*
- * HBMDRAMSIM.H is utilized to bridge the SST-Elements:HBMDRAMSim to DRAMSim
- * with the HBM additions
- *
- */
 
 #ifndef HBMDRAMSIM_H
 #define HBMDRAMSIM_H
 
-#include "DRAMSim.h"
+#include "Callback.h"
+#include <string>
+#include "Stats.h"
 
-#endif // HBMDRAMSIM_H
+using std::string;
+
+namespace HBMDRAMSim
+{
+  class MultiChannelMemorySystem
+  {
+    public:
+      bool addTransaction(bool isWrite, uint64_t addr);
+      void update();
+      void printStats(bool finalStats);
+      bool willAcceptTransaction();
+      bool willAcceptTransaction(uint64_t addr);
+
+      void RegisterCallbacks(TransactionCompleteCB *readDone,
+                             TransactionCompleteCB *writeDone,
+                             void (*reportPower)(double bgpower,
+                                                 double burstpower,
+                                                 double refreshpower,
+                                                 double actprepower));
+
+      // SST Statistics
+      bool getStats( double *stat, DSIM_STAT metric );
+      bool getStats( uint64_t *stat, DSIM_STAT metric );
+  };
+
+  MultiChannelMemorySystem *getMemorySystemInstance(const string &dev, const string &sys, 
+      const string &pwd, unsigned megsOfMemory);
+}
+
+#endif
