+++ MemoryController.cpp
@@ -36,7 +36,7 @@
 
 #define SEQUENTIAL(rank,bank) (rank*NUM_BANKS)+bank
 
-using namespace DRAMSim;
+using namespace HBMDRAMSim;
 
 MemoryController::MemoryController(unsigned sid, unsigned cid, MemorySystem *parent) :
   stackID(sid),
